refactor: Provide request context
The availability of `request` throughout the code is not consistent. Some
methods get `request` passed explicitly, some don't, and some constructs have it
available both as a parameter and as `self.request`. Having `request` available
is essential, especially when trying to implement adapter hooks, though
littering the code base with an explicit `request` is not very great.

With the changes made here, there is always a request available if you need it:
`allauth.core.context.request`.