fix(utils): generate unique username vs cyrillic
For cyrillic names on social accounts - the encode and decode operations result in replacing all chars with empty string. And later in try-except block clean_username does not raise a ValidationError, so empty string username is considered valid. As a result, "user" is returned at first iteration and other txts are not even being processed